Add init mode and use Jinja
This commit uses Jinja to handle creation of two things. First, it
handles creation of the boilerplate code when creating new solutions.
Second, it handles creation of the config file for CPSM. Now, when init
mode is called, it queries the user for several configuration options
and eventually creates a cpsm_config.py file in the current directory.

Along with cpsm.py and the template in cpsm_config_template.j2, the
setup.py and documentation in README were updated.