Allow more templates
Modified CPSM to allow multiple templates for the same filetype. This is
done by giving templates names that are independent of their filetype,
instead of restricting them to only be file extensions. The appropriate
filetype is also included with each template in the config file.

The CPSM script itself was also modified to handle the new templates.

Documentation in README to follow.