Add NPM and devtools
NPM is now used for bringing several dev tools into the project. Mainly,
husky and lint-staged are used to ensure that committed files go through
various checks before they can be added to the project. Meanwhile,
markdown-toc and prettier are used to format markdown files, and yapf
(installed with pip) is used to format Python files.