Post CPSM to PyPI
Several files were modified in order to post CPSM to PyPI. A MANIFEST.in
was added to choose files for distribution, and setup.py was updated to
provide info for PyPI. Version was updated in __init__.py and setup.py;
CPSM is now 1.0.0! Finally, README was updated with new installation
instructions.